app/requirements.txt
fastapi
uvicorn
python-dotenv
requests
groq
langchain-core
langchain-openai
mcp-use
